fix leaderboard
